+++ executables/GLFW/TestGLFW_ParticleRefactor2/src/main.cpp
@@ -178,9 +178,9 @@ protected:
 		chaos::MyGLFW::Window::Finalize();
 	}
 
-	virtual bool Tick(double delta_time) override
+	virtual bool Tick(float delta_time) override
 	{
-		particle_manager->Tick((float)delta_time);
+		particle_manager->Tick(delta_time);
 
 		return true; // no redraw
 	}
