+++ executables/GLFW/TestGLFW_TextSpriteFormat/src/main.cpp
@@ -241,9 +241,9 @@ protected:
 		hints.decorated = 1;
 	}
 
-	virtual bool Tick(double delta_time) override
+	virtual bool Tick(float delta_time) override
 	{
-		particle_manager->Tick((float)delta_time);
+		particle_manager->Tick(delta_time);
 		fps_view_controller.Tick(glfw_window, delta_time);
 
 		return true; // refresh
