+++ executables/LUDUMDARE/Ludum40/src/main.cpp
@@ -109,7 +109,7 @@ protected:
 		hints.decorated = 1;
 	}
 
-	virtual bool Tick(double delta_time) override
+	virtual bool Tick(float delta_time) override
 	{
 		game->Tick(delta_time);
