+++ executables/LUDUMDARE/Ludum41Custom/src/Ludum41CustomGameInstance.cpp
@@ -36,7 +36,7 @@ size_t LudumGameInstance::CanStartChallengeBallIndex(bool going_down) const
 }
 
 
-void LudumGameInstance::TickChallenge(double delta_time)
+void LudumGameInstance::TickChallenge(float delta_time)
 {
 	if (sequence_challenge != nullptr)
 	{
@@ -45,7 +45,7 @@ void LudumGameInstance::TickChallenge(double delta_time)
 	else
 	{
 		// start a challenge (only if one ball is going upward)
-		challenge_timer = std::max(0.0f, challenge_timer - (float)delta_time);
+		challenge_timer = std::max(0.0f, challenge_timer - delta_time);
 		if (challenge_timer <= 0.0f)
 			if (current_background_fillratio > 0.0f) // no challenge when waiting for level change
 				if (CanStartChallengeBallIndex(false) != std::numeric_limits<size_t>::max()) // any ball going up
@@ -64,7 +64,7 @@ static void RotateVelocity(glm::vec2 & src, float angle)
 }
 
 
-void LudumGameInstance::TickBackgroundFillRatio(double delta_time)
+void LudumGameInstance::TickBackgroundFillRatio(float delta_time)
 {
 	LudumGame const* game = GetLudumGame();
 	if (game == nullptr)
@@ -86,13 +86,13 @@ void LudumGameInstance::TickBackgroundFillRatio(double delta_time)
 
 	if (target_fillratio < current_background_fillratio)
 	{
-		current_background_fillratio -= game->background_fillratio_changespeed * (float)delta_time;
+		current_background_fillratio -= game->background_fillratio_changespeed * delta_time;
 		current_background_fillratio = std::max(current_background_fillratio, target_fillratio);
 	}
 
 	// tick for completion timer
 	if (current_background_fillratio <= 0.0f)
-		complete_level_timer += (float)delta_time;
+		complete_level_timer += delta_time;
 }
 
 void LudumGameInstance::FillUniformProvider(chaos::GPUProgramProvider& main_uniform_provider)
@@ -112,7 +112,7 @@ void LudumGameInstance::FillUniformProvider(chaos::GPUProgramProvider& main_unif
 }
 
 
-void LudumGameInstance::TickBallSplit(double delta_time)
+void LudumGameInstance::TickBallSplit(float delta_time)
 {
 	LudumGame const * ludum_game = GetLudumGame();
 
@@ -141,19 +141,19 @@ void LudumGameInstance::TickBallSplit(double delta_time)
 	pending_split_count = 0;
 }
 
-void LudumGameInstance::TickBrickOffset(double delta_time)
+void LudumGameInstance::TickBrickOffset(float delta_time)
 {
 	LudumGame const * ludum_game = GetLudumGame();
 
 	if (target_brick_offset > brick_offset)
 	{
-		brick_offset = brick_offset + (float)delta_time * ludum_game->brick_offset_speed;
+		brick_offset = brick_offset + delta_time * ludum_game->brick_offset_speed;
 		if (brick_offset > target_brick_offset)
 			brick_offset = target_brick_offset;
 	}
 	else if (target_brick_offset < brick_offset)
 	{
-		brick_offset = brick_offset - (float)delta_time * ludum_game->brick_offset_speed;
+		brick_offset = brick_offset - delta_time * ludum_game->brick_offset_speed;
 		if (brick_offset < target_brick_offset)
 			brick_offset = target_brick_offset;
 	}
@@ -254,7 +254,7 @@ bool LudumGameInstance::CanCompleteLevel() const
 	return true;
 }
 
-bool LudumGameInstance::DoTick(double delta_time)
+bool LudumGameInstance::DoTick(float delta_time)
 {
 	if (!death::GameInstance::DoTick(delta_time)) // ticking GameInstance, tick Players ... this is usefull to work with inputs
 		return false;
