+++ executables/LUDUMDARE/Ludum44/src/Ludum44HUD.h
@@ -43,7 +43,7 @@ class GameHUDHealthBarComponent : public death::GameHUDSingleAllocationComponent
 protected:
 
 	/** override */
-	virtual bool DoTick(double delta_time) override;
+	virtual bool DoTick(float delta_time) override;
 };
 
 // ====================================================================
@@ -56,7 +56,7 @@ class GameHUDPowerUpComponent : public death::GameHUDSingleAllocationComponent
 
 protected:
 
-	virtual bool DoTick(double delta_time) override;
+	virtual bool DoTick(float delta_time) override;
 
 protected:
