+++ executables/LUDUMDARE/Ludum44/src/Ludum44LevelInstance.cpp
@@ -34,7 +34,7 @@ void LudumLevelInstance::CreateCameras()
 	}
 }
 
-bool LudumLevelInstance::DoTick(double delta_time)
+bool LudumLevelInstance::DoTick(float delta_time)
 {
 	death::TiledMap::LevelInstance::DoTick(delta_time);
 
@@ -60,7 +60,7 @@ bool LudumLevelInstance::DoTick(double delta_time)
 		return true;
 
 	// compute scroll for both camera and player
-	float scroll_displacement = scroll_factor * ludum_game->scroll_factor * camera_speed * (float)delta_time;
+	float scroll_displacement = scroll_factor * ludum_game->scroll_factor * camera_speed * delta_time;
 	float camera_y = camera_box.position.y + scroll_displacement; // the final wanted camera Y
 
 	// the camera follows the player in X & Y direction
