+++ libraries/death/include/death/Camera.h
@@ -86,7 +86,7 @@ namespace death
 	protected:
 
 		/** override */
-		virtual bool DoTick(double delta_time) override;
+		virtual bool DoTick(float delta_time) override;
 
 	protected:
